Data.Text.ICU.Error

Description

Errors thrown by bindings to the International Components for Unicode (ICU) libraries.

Most ICU functions can throw an ICUError value as an exception. Some can additionally throw a ParseError, if more detailed error information is necessary.

Types

data ICUError

ICU error type. This is an instance of the Exception type class. A value of this type may be thrown as an exception by most ICU functions.

data ParseError

Detailed information about parsing errors. Used by ICU parsing engines that parse long rules, patterns, or programs, where the text being parsed is long enough that more information than an ICUError is needed to localize the error.

Functions

isSuccess :: ICUError -> Bool

Indicate whether the given error code is a success.

isFailure :: ICUError -> Bool

Indicate whether the given error code is a failure.

errorName :: ICUError -> String

Return a string representing the name of the given error code.

isRegexError :: ICUError -> Bool

Warnings

u_USING_FALLBACK_WARNING :: ICUError

u_USING_DEFAULT_WARNING :: ICUError

u_SAFECLONE_ALLOCATED_WARNING :: ICUError

u_STATE_OLD_WARNING :: ICUError

u_STRING_NOT_TERMINATED_WARNING :: ICUError

u_SORT_KEY_TOO_SHORT_WARNING :: ICUError

u_AMBIGUOUS_ALIAS_WARNING :: ICUError

u_DIFFERENT_UCA_VERSION :: ICUError

Errors

u_ILLEGAL_ARGUMENT_ERROR :: ICUError

u_MISSING_RESOURCE_ERROR :: ICUError

u_INVALID_FORMAT_ERROR :: ICUError

u_FILE_ACCESS_ERROR :: ICUError

u_INTERNAL_PROGRAM_ERROR :: ICUError

u_MESSAGE_PARSE_ERROR :: ICUError

u_MEMORY_ALLOCATION_ERROR :: ICUError

u_INDEX_OUTOFBOUNDS_ERROR :: ICUError

u_PARSE_ERROR :: ICUError

u_INVALID_CHAR_FOUND :: ICUError

u_TRUNCATED_CHAR_FOUND :: ICUError

u_ILLEGAL_CHAR_FOUND :: ICUError

u_INVALID_TABLE_FORMAT :: ICUError

u_INVALID_TABLE_FILE :: ICUError

u_BUFFER_OVERFLOW_ERROR :: ICUError

u_UNSUPPORTED_ERROR :: ICUError

u_RESOURCE_TYPE_MISMATCH :: ICUError

u_ILLEGAL_ESCAPE_SEQUENCE :: ICUError

u_UNSUPPORTED_ESCAPE_SEQUENCE :: ICUError

u_NO_SPACE_AVAILABLE :: ICUError

u_CE_NOT_FOUND_ERROR :: ICUError

u_PRIMARY_TOO_LONG_ERROR :: ICUError

u_STATE_TOO_OLD_ERROR :: ICUError

u_TOO_MANY_ALIASES_ERROR :: ICUError

u_ENUM_OUT_OF_SYNC_ERROR :: ICUError

u_INVARIANT_CONVERSION_ERROR :: ICUError

u_INVALID_STATE_ERROR :: ICUError

u_COLLATOR_VERSION_MISMATCH :: ICUError

u_USELESS_COLLATOR_ERROR :: ICUError

u_NO_WRITE_PERMISSION :: ICUError

Transliterator errors

u_BAD_VARIABLE_DEFINITION :: ICUError

u_MALFORMED_RULE :: ICUError

u_MALFORMED_SET :: ICUError

u_MALFORMED_UNICODE_ESCAPE :: ICUError

u_MALFORMED_VARIABLE_DEFINITION :: ICUError

u_MALFORMED_VARIABLE_REFERENCE :: ICUError

u_MISPLACED_CURSOR_OFFSET :: ICUError

u_MISPLACED_QUANTIFIER :: ICUError

u_MISSING_OPERATOR :: ICUError

u_MULTIPLE_ANTE_CONTEXTS :: ICUError

u_MULTIPLE_CURSORS :: ICUError

u_MULTIPLE_POST_CONTEXTS :: ICUError

u_TRAILING_BACKSLASH :: ICUError

u_UNDEFINED_SEGMENT_REFERENCE :: ICUError

u_UNDEFINED_VARIABLE :: ICUError

u_UNQUOTED_SPECIAL :: ICUError

u_UNTERMINATED_QUOTE :: ICUError

u_RULE_MASK_ERROR :: ICUError

u_MISPLACED_COMPOUND_FILTER :: ICUError

u_MULTIPLE_COMPOUND_FILTERS :: ICUError

u_INVALID_RBT_SYNTAX :: ICUError

u_MALFORMED_PRAGMA :: ICUError

u_UNCLOSED_SEGMENT :: ICUError

u_VARIABLE_RANGE_EXHAUSTED :: ICUError

u_VARIABLE_RANGE_OVERLAP :: ICUError

u_ILLEGAL_CHARACTER :: ICUError

u_INTERNAL_TRANSLITERATOR_ERROR :: ICUError

u_INVALID_ID :: ICUError

u_INVALID_FUNCTION :: ICUError

Formatting API parsing errors

u_UNEXPECTED_TOKEN :: ICUError

u_MULTIPLE_DECIMAL_SEPARATORS :: ICUError

u_MULTIPLE_EXPONENTIAL_SYMBOLS :: ICUError

u_MALFORMED_EXPONENTIAL_PATTERN :: ICUError

u_MULTIPLE_PERCENT_SYMBOLS :: ICUError

u_MULTIPLE_PERMILL_SYMBOLS :: ICUError

u_MULTIPLE_PAD_SPECIFIERS :: ICUError

u_PATTERN_SYNTAX_ERROR :: ICUError

u_ILLEGAL_PAD_POSITION :: ICUError

u_UNMATCHED_BRACES :: ICUError

u_ARGUMENT_TYPE_MISMATCH :: ICUError

u_DUPLICATE_KEYWORD :: ICUError

u_UNDEFINED_KEYWORD :: ICUError

u_DEFAULT_KEYWORD_MISSING :: ICUError

Break iterator errors

u_BRK_INTERNAL_ERROR :: ICUError

u_BRK_HEX_DIGITS_EXPECTED :: ICUError

u_BRK_SEMICOLON_EXPECTED :: ICUError

u_BRK_RULE_SYNTAX :: ICUError

u_BRK_UNCLOSED_SET :: ICUError

u_BRK_ASSIGN_ERROR :: ICUError

u_BRK_VARIABLE_REDFINITION :: ICUError

u_BRK_MISMATCHED_PAREN :: ICUError

u_BRK_NEW_LINE_IN_QUOTED_STRING :: ICUError

u_BRK_UNDEFINED_VARIABLE :: ICUError

u_BRK_INIT_ERROR :: ICUError

u_BRK_RULE_EMPTY_SET :: ICUError

u_BRK_UNRECOGNIZED_OPTION :: ICUError

u_BRK_MALFORMED_RULE_TAG :: ICUError

Regular expression errors

u_REGEX_INTERNAL_ERROR :: ICUError

u_REGEX_RULE_SYNTAX :: ICUError

u_REGEX_INVALID_STATE :: ICUError

u_REGEX_BAD_ESCAPE_SEQUENCE :: ICUError

u_REGEX_PROPERTY_SYNTAX :: ICUError

u_REGEX_UNIMPLEMENTED :: ICUError

u_REGEX_MISMATCHED_PAREN :: ICUError

u_REGEX_NUMBER_TOO_BIG :: ICUError

u_REGEX_BAD_INTERVAL :: ICUError

u_REGEX_MAX_LT_MIN :: ICUError

u_REGEX_INVALID_BACK_REF :: ICUError

u_REGEX_INVALID_FLAG :: ICUError

u_REGEX_SET_CONTAINS_STRING :: ICUError

u_REGEX_OCTAL_TOO_BIG :: ICUError

u_REGEX_INVALID_RANGE :: ICUError

u_REGEX_STACK_OVERFLOW :: ICUError

u_REGEX_TIME_OUT :: ICUError

u_REGEX_STOPPED_BY_CALLER :: ICUError

IDNA errors

u_IDNA_PROHIBITED_ERROR :: ICUError

u_IDNA_UNASSIGNED_ERROR :: ICUError

u_IDNA_CHECK_BIDI_ERROR :: ICUError

u_IDNA_STD3_ASCII_RULES_ERROR :: ICUError

u_IDNA_ACE_PREFIX_ERROR :: ICUError

u_IDNA_VERIFICATION_ERROR :: ICUError

u_IDNA_LABEL_TOO_LONG_ERROR :: ICUError

u_IDNA_ZERO_LENGTH_LABEL_ERROR :: ICUError

u_IDNA_DOMAIN_NAME_TOO_LONG_ERROR :: ICUError