Data.Text.ICU.Types

Description

Types for use when manipulating Unicode text, using the bindings to the International Components for Unicode (ICU) libraries.

Widely used types

data LocaleName

The name of a locale.

Constructors

Root
The root locale. For a description of resource bundles and the root resource, see http://userguide.icu-project.org/locale/resources.

Locale String
A specific locale.

Current
The program's current locale.

data ParseError

Detailed information about parsing errors. Used by ICU parsing engines that parse long rules, patterns, or programs, where the text being parsed is long enough that more information than an ICUError is needed to localize the error.